feat(q-flux): Wire PeerTracker, OCSP stapling, TLS drain, H2 proxy rewrite
PeerTracker integration (proxy.rs, worker.rs):
- Wire PeerTracker into WebSocket upgrade path with libp2p handshake detection
- Deny connections from banned/circuit-broken peers via should_allow_peer()
- Track peer connections (conn_opened/conn_closed) for scoring
- Create PeerTracker with known bootstrap/supernode peer IDs
- Add periodic cleanup task (60s interval, 300s stale threshold)

OCSP stapling + TLS drain (acceptor.rs, config.rs, admin.rs):
- Add OCSP stapling via rustls with_single_cert_with_ocsp()
- Add ocsp_staple config field for DER-encoded OCSP response path
- Restructure SharedTlsConfig with drain notification (tokio::sync::Notify)
- Add drain_timeout_secs config for graceful TLS reload drain
- Export tls_reload_count in admin status and Prometheus metrics

HTTP/2 proxy rewrite (h2_proxy.rs):
- Replace placeholder with real hyper HTTP/2 server using service_fn
- Add per-request routing: static files, CORS preflight, upstream proxy
- Add H2Metrics with static LazyLock for stream/request/error counters
- Add h2_prometheus_export() for Prometheus scraping
- Configure initial_stream_window_size and max_frame_size

+++ crates/q-flux/src/libp2p_aware.rs
@@ -636,7 +636,7 @@ impl BandwidthLimiter {
         let elapsed_us = now_us.saturating_sub(last);
         if elapsed_us > 1000 {
             // Refill every 1ms minimum
-            let new_tokens = (elapsed_us * bucket.rate_bytes_per_sec) / 1_000_000;
+            let new_tokens = (elapsed_us as u128 * bucket.rate_bytes_per_sec as u128 / 1_000_000) as u64;
             if new_tokens > 0
                 && bucket
                     .last_refill_us
